refactor: extract api_token service from route handler (#96 step 3)
Move Request/Response types, private helpers (generate_raw_token, hash_token),
and usecase logic for create/list/revoke API tokens from routes/api_token.rs
to services/api_token.rs. Handlers now only parse extractors and delegate to
service functions.

No OpenAPI snapshot changes. All 15 api_token tests pass.
